news-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
httpx
feedparser
transformers
torch
python-dotenv
tf-keras
pytest
